+++ src/devices/generic_device.rs
@@ -1,10 +1,6 @@
-use std::fmt;
-
-use serde::Serialize;
-
 use crate::devices::TapoDeviceExt;
 
 /// Basic functionality of all [Tapo devices](https://www.tapo.com/uk/).
-#[derive(fmt::Debug, Default, Serialize)]
+#[derive(Debug)]
 pub struct GenericDevice;
 impl TapoDeviceExt for GenericDevice {}