+++ src/devices/p100.rs
@@ -1,10 +1,6 @@
-use std::fmt;
-
-use serde::Serialize;
-
 use crate::devices::TapoDeviceExt;
 
 /// [Tapo P100](https://www.tapo.com/uk/search/?q=P100) devices.
-#[derive(fmt::Debug, Default, Serialize)]
+#[derive(Debug)]
 pub struct P100;
 impl TapoDeviceExt for P100 {}