+++ src/devices/p110.rs
@@ -1,10 +1,6 @@
-use std::fmt;
-
-use serde::Serialize;
-
 use crate::devices::TapoDeviceExt;
 
 /// [Tapo P110](https://www.tapo.com/uk/search/?q=P110) devices.
-#[derive(fmt::Debug, Default, Serialize)]
+#[derive(Debug)]
 pub struct P110;
 impl TapoDeviceExt for P110 {}