+++ src/devices/tapo_device_ext.rs
@@ -1,6 +1,2 @@
-use std::fmt;
-
-use serde::Serialize;
-
 /// Implemented by all Tapo devices.
-pub trait TapoDeviceExt: fmt::Debug + Default + Serialize {}
+pub trait TapoDeviceExt: std::fmt::Debug {}